deleting the old give page, updating test, deleting recoil stuff
