--- src/pages/AllocationPage/BalanceDescription.tsx
@@ -1,12 +1,0 @@
-import { styled } from '@stitches/react';
-
-const BalanceDescription = styled('p', {
-  margin: 0,
-  fontSize: 20,
-  fontWeight: 300,
-  '&:first-of-type': {
-    fontWeight: 500,
-  },
-});
-
-export default BalanceDescription;
