--- src/pages/AllocationPage/SaveButtonContainer.tsx
@@ -1,15 +1,0 @@
-import { styled } from '@stitches/react';
-
-const SaveButtonContainer = styled('div', {
-  position: 'fixed',
-  bottom: 0,
-  marginBottom: 53,
-  marginLeft: 'auto',
-  marginRight: 'auto',
-  display: 'flex',
-  flexDirection: 'row',
-  justifyContent: 'center',
-  alignItems: 'center',
-});
-
-export default SaveButtonContainer;
