--- src/pages/AllocationPage/index.ts
@@ -1,3 +1,0 @@
-import AllocationPage from './AllocationPage';
-
-export default AllocationPage;
