+++ src/recoilState/index.ts
@@ -2,4 +2,3 @@ export * from './app';
 export * from './db';
 export * from './map';
 export * from './ui';
-export * from './allocation';
